Moved CSS file; made calendar smaller, added a button for the Nutrition Report
