+++ src/pages/calendar/Calendar.jsx
@@ -3,7 +3,7 @@ import Calendar from "react-calendar";
 import Modal from "react-modal";
 import { eachDayOfInterval, startOfDay, isSameDay } from "date-fns";
 import MealForm from "./components/MealForm.jsx";
-import "../../css/calendarStyle.css";
+import "./calendarStyle.css";
 import "react-calendar/dist/Calendar.css";
 import FirestoreService from "../../firebase/FirebaseService.js";
 import FirestoreListener from "../../firebase/FirestoreListener.js";
@@ -32,7 +32,7 @@ const MyCalendar = () => {
       (docs) => {
         const plans = docs;
         setPlans(plans);
-      },
+      }
     );
 
     return unsubscribeFromPlans;
@@ -81,7 +81,7 @@ const MyCalendar = () => {
         `Users/${user.uid}/Plans/`,
         planDate,
         existingPlan,
-        "plan",
+        "plan"
       ).catch((error) => console.error("Error updating plan: ", error));
     }
     //if there's no existing plan, create a new plan
@@ -103,7 +103,7 @@ const MyCalendar = () => {
         `Users/${user.uid}/Plans/`,
         planDate,
         newPlan,
-        "plan",
+        "plan"
       ).catch((error) => console.error("Error creating plan: ", error));
     }
   };
@@ -114,19 +114,19 @@ const MyCalendar = () => {
     //if the plan exists, remove the selected meal
     if (planToUpdate) {
       const updatedMeals = planToUpdate.meals.filter(
-        (meal, index) => index !== mealIndex,
+        (meal, index) => index !== mealIndex
       );
       //updating plan with new meal list
       const updatedPlan = { ...planToUpdate, meals: updatedMeals };
       //updating plan in the local state
       setPlans(
-        plans.map((plan) => (plan.date === planDate ? updatedPlan : plan)),
+        plans.map((plan) => (plan.date === planDate ? updatedPlan : plan))
       );
       //updating firestore
       FirestoreService.updateDocument(
         `Users/${user.uid}/Plans/`,
         planDate,
-        updatedPlan,
+        updatedPlan
       ).catch((error) => console.error("Error removing meal: ", error));
     }
   };
@@ -148,12 +148,12 @@ const MyCalendar = () => {
         onClickDay={onClickDay}
         tileContent={({ date, view }) => {
           const dayPlans = plans.filter(
-            (plan) => plan.date === date.toISOString().split("T")[0],
+            (plan) => plan.date === date.toISOString().split("T")[0]
           );
           //finding out how many meals are planned for that day
           const totalMeals = dayPlans.reduce(
             (sum, plan) => sum + plan.meals.length,
-            0,
+            0
           );
           return (
             <div>
@@ -181,46 +181,51 @@ const MyCalendar = () => {
           }
         }}
       />
-      <div className="selected-day">
-        <span className="date-display">
-          {selectedDay.toLocaleString("en-US", {
-            weekday: "long",
-            month: "short",
-            day: "numeric",
-          })}
-        </span>{" "}
-        <br />
-        <Modal
-          ariaHideApp={false}
-          isOpen={isModalOpen}
-          onRequestClose={closeModal}
-        >
-          <MealForm closeModal={closeModal} addPlan={addPlan} />
-        </Modal>
-        <br />
-        {plans
-          //filtering plans by selected day to display them
-          .filter(
-            (plan) => plan.date === selectedDay.toISOString().split("T")[0],
-          )
-          .map((plan, index) =>
-            plan.meals.map((meal, mealIndex) => (
-              //for each entry, create a div displaying the meal's name
-              <div key={mealIndex} className="meal-tile rounded">
-                <h6>{meal.name}</h6>
-                <button
-                  type="button"
-                  className="rm-meal-btn"
-                  onClick={() => removePlan(plan.date, mealIndex)}
-                >
-                  Remove
-                </button>
-              </div>
-            )),
-          )}
-        <button className="add-meal-btn" onClick={openModal}>
-          Add Meal
-        </button>
+      <div id="calendar-sidebar">
+        <div id="nutrition-launcher">
+          <button id="nutrition-button">Generate Nutrition Report</button>
+        </div>
+        <div className="selected-day">
+          <span className="date-display">
+            {selectedDay.toLocaleString("en-US", {
+              weekday: "long",
+              month: "short",
+              day: "numeric",
+            })}
+          </span>{" "}
+          <br />
+          <Modal
+            ariaHideApp={false}
+            isOpen={isModalOpen}
+            onRequestClose={closeModal}
+          >
+            <MealForm closeModal={closeModal} addPlan={addPlan} />
+          </Modal>
+          <br />
+          {plans
+            //filtering plans by selected day to display them
+            .filter(
+              (plan) => plan.date === selectedDay.toISOString().split("T")[0]
+            )
+            .map((plan, index) =>
+              plan.meals.map((meal, mealIndex) => (
+                //for each entry, create a div displaying the meal's name
+                <div key={mealIndex} className="meal-tile rounded">
+                  <h6>{meal.name}</h6>
+                  <button
+                    type="button"
+                    className="rm-meal-btn"
+                    onClick={() => removePlan(plan.date, mealIndex)}
+                  >
+                    Remove
+                  </button>
+                </div>
+              ))
+            )}
+          <button className="add-meal-btn" onClick={openModal}>
+            Add Meal
+          </button>
+        </div>
       </div>
     </div>
   );
