+++ src/pages/calendar/calendarStyle.css
@@ -9,6 +9,24 @@
 .calendar-container {
   display: flex;
   justify-content: center;
+  overflow: hidden;
+}
+
+#calendar-sidebar {
+  display: flex;
+  flex-direction: column;
+  width: 25%;
+  /* max-width: 300px; */
+}
+
+#nutrition-launcher {
+  width: 100%;
+  height: 10%;
+}
+
+#nutrition-button {
+  width: 100%;
+  height: 100%;
 }
 
 .selected-day {
@@ -17,7 +35,8 @@
   color: white;
   padding-left: 5%;
   padding-right: 5%;
-  width: 30%;
+  width: 100%;
+  height: 90%;
   overflow: hidden;
   text-overflow: ellipsis;
 }
@@ -57,9 +76,8 @@
 }
 
 .date-display {
-  font-size: 1.1em;
+  font-size: 1em;
   font-weight: bold;
-  /* color: #8b0000; */
   color: black;
 }
 
@@ -69,13 +87,12 @@
 
 .meal-tile {
   background-color: #ff8c8c;
-  margin-bottom: 3%;
+  margin-bottom: 2%;
 }
 
-/* react classes have to be overridden with !important */
-
 .react-calendar {
-  width: 100% !important;
+  width: 70% !important;
+  /* max-width: 500px; */
   background: #f0f0f0 !important;
   border: none !important;
   font-family: "Roboto", sans-serif !important;
@@ -88,14 +105,14 @@
 }
 
 .react-calendar__month-view__weekdays__weekday {
-  padding: 1.3em !important;
+  padding: 0.8em !important;
   color: #999999 !important;
 }
 
 .react-calendar__month-view__days__day {
-  padding: 1.5em !important;
+  padding: 1em !important;
   text-align: center !important;
-  font-size: 1em !important;
+  font-size: 0.9em !important;
 }
 
 .react-calendar__month-view__days__day--weekend {
@@ -128,7 +145,7 @@
 
 .react-calendar__navigation__label {
   font-family: "Arial", sans-serif !important;
-  font-size: 2em !important;
+  font-size: 1.5em !important;
 }
 
 .selected-dates {
@@ -139,21 +156,21 @@
   display: flex;
   flex-direction: column;
   width: 90%;
-  max-width: 300px;
+  /* max-width: 300px; */
   margin: 0 auto;
   text-align: center;
 }
 
 .meal-form-container label {
-  margin-bottom: 1em;
+  margin-bottom: 0.8em;
   color: #570000;
-  font-size: 1em;
+  font-size: 0.9em;
 }
 
 .meal-form-container input[type="text"],
 .meal-form-container input[type="checkbox"] {
   width: 100%;
-  padding: 1em;
+  padding: 0.8em;
   margin-bottom: 1em;
   border: 1px solid #ccc;
   border-radius: 0.25em;
@@ -162,14 +179,14 @@
 
 .meal-form-container input[type="submit"],
 .meal-form-container button {
-  padding: 1em;
+  padding: 0.8em;
   border: none;
   border-radius: 0.75em;
   background-color: #ff7f7f;
   color: white;
   cursor: pointer;
   transition-duration: 0.4s;
-  font-size: 1em;
+  font-size: 0.9em;
   text-align: center;
 }
