+++ src/pages/calendar/components/MealForm.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { useForm } from "react-hook-form";
-import "../../../css/calendarStyle.css";
+import "../calendarStyle.css";
 import FirestoreListener from "../../../firebase/FirestoreListener.js";
 import { useAuth } from "../../../utils/AuthContext.js";
